Fix CppCheck in rschatitems.h
/libretroshare/src/chat/rschatitems.h:114: warning:
Cppcheck(uninitMemberVar): Member variable 'RsChatMsgItem::chatFlags' is
not initialized in the constructor.
/libretroshare/src/chat/rschatitems.h:114: warning:
Cppcheck(uninitMemberVar): Member variable 'RsChatMsgItem::sendTime' is
not initialized in the constructor.
/libretroshare/src/chat/rschatitems.h:114: warning:
Cppcheck(uninitMemberVar): Member variable 'RsChatMsgItem::recvTime' is
not initialized in the constructor.
/libretroshare/src/chat/rschatitems.h:115: warning:
Cppcheck(uninitMemberVar): Member variable 'RsChatMsgItem::chatFlags' is
not initialized in the constructor.
/libretroshare/src/chat/rschatitems.h:115: warning:
Cppcheck(uninitMemberVar): Member variable 'RsChatMsgItem::sendTime' is
not initialized in the constructor.
/libretroshare/src/chat/rschatitems.h:115: warning:
Cppcheck(uninitMemberVar): Member variable 'RsChatMsgItem::recvTime' is
not initialized in the constructor.
/libretroshare/src/chat/rschatitems.h:161: warning:
Cppcheck(uninitMemberVar): Member variable
'RsChatLobbyMsgItem::parent_msg_id' is not initialized in the
constructor.
/libretroshare/src/chat/rschatitems.h:177: warning:
Cppcheck(uninitMemberVar): Member variable
'RsChatLobbyEventItem::event_type' is not initialized in the
constructor.
/libretroshare/src/chat/rschatitems.h:177: warning:
Cppcheck(uninitMemberVar): Member variable
'RsChatLobbyEventItem::sendTime' is not initialized in the constructor.
/libretroshare/src/chat/rschatitems.h:226: warning:
Cppcheck(uninitMemberVar): Member variable
'RsChatLobbyUnsubscribeItem::lobby_id' is not initialized in the
constructor.
/libretroshare/src/chat/rschatitems.h:238: warning:
Cppcheck(uninitMemberVar): Member variable
'RsChatLobbyConnectChallengeItem::challenge_code' is not initialized in
the constructor.
/libretroshare/src/chat/rschatitems.h:251: warning:
Cppcheck(uninitMemberVar): Member variable
'RsChatLobbyInviteItem_Deprecated::lobby_id' is not initialized in the
constructor.
/libretroshare/src/chat/rschatitems.h:265: warning:
Cppcheck(uninitMemberVar): Member variable
'RsChatLobbyInviteItem::lobby_id' is not initialized in the constructor.
/libretroshare/src/chat/rschatitems.h:282: warning:
Cppcheck(uninitMemberVar): Member variable
'RsPrivateChatMsgConfigItem::chatFlags' is not initialized in the
constructor.
/libretroshare/src/chat/rschatitems.h:282: warning:
Cppcheck(uninitMemberVar): Member variable
'RsPrivateChatMsgConfigItem::configFlags' is not initialized in the
constructor.
/libretroshare/src/chat/rschatitems.h:282: warning:
Cppcheck(uninitMemberVar): Member variable
'RsPrivateChatMsgConfigItem::sendTime' is not initialized in the
constructor.
/libretroshare/src/chat/rschatitems.h:282: warning:
Cppcheck(uninitMemberVar): Member variable
'RsPrivateChatMsgConfigItem::recvTime' is not initialized in the
constructor.
/libretroshare/src/chat/rschatitems.h:307: warning:
Cppcheck(uninitMemberVar): Member variable
'RsChatLobbyConfigItem::flags' is not initialized in the constructor.
/libretroshare/src/chat/rschatitems.h:324: warning:
Cppcheck(uninitMemberVar): Member variable 'RsChatStatusItem::flags' is
not initialized in the constructor.
/libretroshare/src/chat/rschatitems.h:339: warning:
Cppcheck(uninitMemberVar): Member variable
'RsChatAvatarItem::image_size' is not initialized in the constructor.
/libretroshare/src/chat/rschatitems.h:339: warning:
Cppcheck(uninitMemberVar): Member variable
'RsChatAvatarItem::image_data' is not initialized in the constructor./
home/phenom/GIT/RetroShare/trunk/libretroshare/src/chat/rschatitems.h:
96: warning: Cppcheck(noExplicitConstructor): Class 'RsChatItem' has a
constructor with 1 argument that is not explicit.
/libretroshare/src/chat/rschatitems.h:115: warning:
Cppcheck(noExplicitConstructor): Class 'RsChatMsgItem' has a constructor
with 1 argument that is not explicit.
/libretroshare/src/chat/rschatitems.h:361: warning:
Cppcheck(noExplicitConstructor): Struct 'RsChatSerialiser' has a
constructor with 1 argument that is not explicit.