Fix CppCheck in rsfilelistitems.h
/libretroshare/src/file_sharing/rsfilelistitems.h:78: warning:
Cppcheck(uninitMemberVar): Member variable
'RsFileListsSyncRequestItem::flags' is not initialized in the
constructor.
/libretroshare/src/file_sharing/rsfilelistitems.h:78: warning:
Cppcheck(uninitMemberVar): Member variable
'RsFileListsSyncRequestItem::last_known_recurs_modf_TS' is not
initialized in the constructor.
/libretroshare/src/file_sharing/rsfilelistitems.h:78: warning:
Cppcheck(uninitMemberVar): Member variable
'RsFileListsSyncRequestItem::request_id' is not initialized in the
constructor.
/libretroshare/src/file_sharing/rsfilelistitems.h:94: warning:
Cppcheck(uninitMemberVar): Member variable
'RsFileListsSyncResponseItem::flags' is not initialized in the
constructor.
/libretroshare/src/file_sharing/rsfilelistitems.h:94: warning:
Cppcheck(uninitMemberVar): Member variable
'RsFileListsSyncResponseItem::last_known_recurs_modf_TS' is not
initialized in the constructor.
/libretroshare/src/file_sharing/rsfilelistitems.h:94: warning:
Cppcheck(uninitMemberVar): Member variable
'RsFileListsSyncResponseItem::request_id' is not initialized in the
constructor.
/libretroshare/src/file_sharing/rsfilelistitems.h:52: warning:
Cppcheck(noExplicitConstructor): Class 'RsFileListsItem' has a
constructor with 1 argument that is not explicit.