Fix CppCheck in grouteritems.h
/libretroshare/src/grouter/grouteritems.h:93: warning:
Cppcheck(uninitMemberVar): Member variable
'RsGRouterAbstractMsgItem::flags' is not initialized in the constructor.
/libretroshare/src/grouter/grouteritems.h:106: warning:
Cppcheck(uninitMemberVar): Member variable
'RsGRouterGenericDataItem::data_size' is not initialized in the
constructor.
/libretroshare/src/grouter/grouteritems.h:106: warning:
Cppcheck(uninitMemberVar): Member variable
'RsGRouterGenericDataItem::data_bytes' is not initialized in the
constructor.
/libretroshare/src/grouter/grouteritems.h:106: warning:
Cppcheck(uninitMemberVar): Member variable
'RsGRouterGenericDataItem::duplication_factor' is not initialized in the
constructor.
/libretroshare/src/grouter/grouteritems.h:159: warning:
Cppcheck(uninitMemberVar): Member variable
'RsGRouterTransactionChunkItem::chunk_start' is not initialized in the
constructor.
/libretroshare/src/grouter/grouteritems.h:159: warning:
Cppcheck(uninitMemberVar): Member variable
'RsGRouterTransactionChunkItem::chunk_size' is not initialized in the
constructor.
/libretroshare/src/grouter/grouteritems.h:159: warning:
Cppcheck(uninitMemberVar): Member variable
'RsGRouterTransactionChunkItem::total_size' is not initialized in the
constructor.
/libretroshare/src/grouter/grouteritems.h:159: warning:
Cppcheck(uninitMemberVar): Member variable
'RsGRouterTransactionChunkItem::chunk_data' is not initialized in the
constructor.
/libretroshare/src/grouter/grouteritems.h:220: warning:
Cppcheck(uninitMemberVar): Member variable
'RsGRouterMatrixTrackItem::time_stamp' is not initialized in the
constructor.
/libretroshare/src/grouter/grouteritems.h:62: warning:
Cppcheck(noExplicitConstructor): Class 'RsGRouterItem' has a constructor
with 1 argument that is not explicit.
/libretroshare/src/grouter/grouteritems.h:93: warning:
Cppcheck(noExplicitConstructor): Class 'RsGRouterAbstractMsgItem' has a
constructor with 1 argument that is not explicit.
/libretroshare/src/grouter/grouteritems.h:147: warning:
Cppcheck(noExplicitConstructor): Class 'RsGRouterTransactionItem' has a
constructor with 1 argument that is not explicit.
/libretroshare/src/grouter/grouteritems.h:274: warning:
Cppcheck(noExplicitConstructor): Class 'RsGRouterSerialiser' has a
constructor with 1 argument that is not explicit.