Fix CppCheck in rsgds.h
/libretroshare/src/gxs/rsgds.h:86: warning: Cppcheck(uninitMemberVar):
Member variable 'RsGroupNetworkStats::mSuppliers' is not initialized in
the constructor.
/libretroshare/src/gxs/rsgds.h:86: warning: Cppcheck(uninitMemberVar):
Member variable 'RsGroupNetworkStats::mGrpAutoSync' is not initialized
in the constructor.
/libretroshare/src/gxs/rsgds.h:86: warning: Cppcheck(uninitMemberVar):
Member variable 'RsGroupNetworkStats::mAllowMsgSync' is not initialized
in the constructor.
/libretroshare/src/gxs/rsgds.h:86: warning: Cppcheck(uninitMemberVar):
Member variable 'RsGroupNetworkStats::mLastGroupModificationTS' is not
initialized in the constructor.
/libretroshare/src/gxs/rsgds.h:55: warning:
Cppcheck(copyCtorAndEqOperator): The class 'MsgLocMetaData' has 'copy
constructor' but lack of 'operator='.
/libretroshare/src/gxs/rsgds.h:68: warning:
Cppcheck(copyCtorAndEqOperator): The class 'GrpLocMetaData' has 'copy
constructor' but lack of 'operator='.