Fix CppCheck in rsgxsnetservice.cc
/libretroshare/src/gxs/rsgxsnetservice.cc:3593: warning:
Cppcheck(unusedVariable): Unused variable: decrypted_items
/libretroshare/src/gxs/rsgxsnetservice.cc:1062: warning:
Cppcheck(noExplicitConstructor): Struct 'GrpFragCollate' has a
constructor with 1 argument that is not explicit.
/libretroshare/src/gxs/rsgxsnetservice.cc:1322: warning:
Cppcheck(noExplicitConstructor): Struct 'MsgFragCollate' has a
constructor with 1 argument that is not explicit.
/libretroshare/src/gxs/rsgxsnetservice.cc:2985: warning:
Cppcheck(nullPointerRedundantCheck): Either the condition 'item' is
redundant or there is possible null pointer dereference: item.
/libretroshare/src/gxs/rsgxsnetservice.cc:868: warning:
Cppcheck(unreadVariable): Variable 'it' is assigned a value that is
never used.
/libretroshare/src/gxs/rsgxsnetservice.cc:868: warning:
Cppcheck(unreadVariable): Variable 'it' is assigned a value that is
never used.
/libretroshare/src/gxs/rsgxsnetservice.cc:3094: warning:
Cppcheck(nullPointerRedundantCheck): Either the condition 'item' is
redundant or there is possible null pointer dereference: item.
/libretroshare/src/gxs/rsgxsnetservice.cc:4702: warning:
Cppcheck(nullPointerRedundantCheck): Either the condition '!item' is
redundant or there is possible null pointer dereference: item.
/libretroshare/src/gxs/rsgxsnetservice.cc:2751: warning:
Cppcheck(unreadVariable): Variable 'cutoff' is assigned a value that is
never used.