Fix CppCheck in rsgxstunnelitems.h
/libretroshare/src/gxstunnel/rsgxstunnelitems.h:75: warning:
Cppcheck(uninitMemberVar): Member variable 'RsGxsTunnelDataItem::flags'
is not initialized in the constructor.
/libretroshare/src/gxstunnel/rsgxstunnelitems.h:76: warning:
Cppcheck(uninitMemberVar): Member variable
'RsGxsTunnelDataItem::unique_item_counter' is not initialized in the
constructor.
/libretroshare/src/gxstunnel/rsgxstunnelitems.h:76: warning:
Cppcheck(uninitMemberVar): Member variable 'RsGxsTunnelDataItem::flags'
is not initialized in the constructor.
/libretroshare/src/gxstunnel/rsgxstunnelitems.h:76: warning:
Cppcheck(uninitMemberVar): Member variable
'RsGxsTunnelDataItem::service_id' is not initialized in the constructor.
/libretroshare/src/gxstunnel/rsgxstunnelitems.h:111: warning:
Cppcheck(uninitMemberVar): Member variable
'RsGxsTunnelDataAckItem::unique_item_counter' is not initialized in the
constructor.
/libretroshare/src/gxstunnel/rsgxstunnelitems.h:128: warning:
Cppcheck(uninitMemberVar): Member variable
'RsGxsTunnelDHPublicKeyItem::public_key' is not initialized in the
constructor.
/libretroshare/src/gxstunnel/rsgxstunnelitems.h:145: warning:
Cppcheck(operatorEqVarError): Member variable
'RsGxsTunnelDHPublicKeyItem::public_key' is not assigned a value in
'RsGxsTunnelDHPublicKeyItem::operator='.
/libretroshare/src/gxstunnel/rsgxstunnelitems.h:56: warning:
Cppcheck(noExplicitConstructor): Class 'RsGxsTunnelItem' has a
constructor with 1 argument that is not explicit.
/libretroshare/src/gxstunnel/rsgxstunnelitems.h:76: warning:
Cppcheck(noExplicitConstructor): Class 'RsGxsTunnelDataItem' has a
constructor with 1 argument that is not explicit.