Fix CppCheck in rsdiscovery2items.h
/libretroshare/src/rsitems/rsdiscovery2items.h:61: warning:
Cppcheck(uninitMemberVar): Member variable 'RsDiscPgpListItem::mode' is
not initialized in the constructor.
/libretroshare/src/rsitems/rsdiscovery2items.h:102: warning:
Cppcheck(uninitMemberVar): Member variable 'RsDiscContactItem::netMode'
is not initialized in the constructor.
/libretroshare/src/rsitems/rsdiscovery2items.h:102: warning:
Cppcheck(uninitMemberVar): Member variable 'RsDiscContactItem::vs_disc'
is not initialized in the constructor.
/libretroshare/src/rsitems/rsdiscovery2items.h:102: warning:
Cppcheck(uninitMemberVar): Member variable 'RsDiscContactItem::vs_dht'
is not initialized in the constructor.
/libretroshare/src/rsitems/rsdiscovery2items.h:102: warning:
Cppcheck(uninitMemberVar): Member variable
'RsDiscContactItem::lastContact' is not initialized in the constructor.
/libretroshare/src/rsitems/rsdiscovery2items.h:102: warning:
Cppcheck(uninitMemberVar): Member variable 'RsDiscContactItem::isHidden'
is not initialized in the constructor.
/libretroshare/src/rsitems/rsdiscovery2items.h:102: warning:
Cppcheck(uninitMemberVar): Member variable
'RsDiscContactItem::hiddenPort' is not initialized in the constructor.
/libretroshare/src/rsitems/rsdiscovery2items.h:49: warning:
Cppcheck(noExplicitConstructor): Class 'RsDiscItem' has a constructor
with 1 argument that is not explicit.