Fix CppCheck in rsfiletransferitems.h
/libretroshare/src/rsitems/rsfiletransferitems.h:68: warning:
Cppcheck(uninitMemberVar): Member variable
'RsFileTransferDataRequestItem::fileoffset' is not initialized in the
constructor.
/libretroshare/src/rsitems/rsfiletransferitems.h:68: warning:
Cppcheck(uninitMemberVar): Member variable
'RsFileTransferDataRequestItem::chunksize' is not initialized in the
constructor.
/libretroshare/src/rsitems/rsfiletransferitems.h:107: warning:
Cppcheck(uninitMemberVar): Member variable
'RsFileTransferChunkMapRequestItem::is_client' is not initialized in the
constructor.
/libretroshare/src/rsitems/rsfiletransferitems.h:125: warning:
Cppcheck(uninitMemberVar): Member variable
'RsFileTransferChunkMapItem::is_client' is not initialized in the
constructor.
/libretroshare/src/rsitems/rsfiletransferitems.h:145: warning:
Cppcheck(uninitMemberVar): Member variable
'RsFileTransferSingleChunkCrcRequestItem::chunk_number' is not
initialized in the constructor.
/libretroshare/src/rsitems/rsfiletransferitems.h:163: warning:
Cppcheck(uninitMemberVar): Member variable
'RsFileTransferSingleChunkCrcItem::chunk_number' is not initialized in
the constructor.
/libretroshare/src/rsitems/rsfiletransferitems.h:58: warning:
Cppcheck(noExplicitConstructor): Class 'RsFileTransferItem' has a
constructor with 1 argument that is not explicit.