Fix Cppcheck in rsgxsreputationitems.h
/libretroshare/src/rsitems/rsgxsreputationitems.h:67: warning:
Cppcheck(uninitMemberVar): Member variable
'RsGxsReputationConfigItem::mLatestUpdate' is not initialized in the
constructor.
/libretroshare/src/rsitems/rsgxsreputationitems.h:67: warning:
Cppcheck(uninitMemberVar): Member variable
'RsGxsReputationConfigItem::mLastQuery' is not initialized in the
constructor.
/libretroshare/src/rsitems/rsgxsreputationitems.h:129: warning:
Cppcheck(uninitMemberVar): Member variable
'RsGxsReputationBannedNodeSetItem::mLastActivityTS' is not initialized
in the constructor.
/libretroshare/src/rsitems/rsgxsreputationitems.h:144: warning:
Cppcheck(uninitMemberVar): Member variable
'RsGxsReputationUpdateItem::mLatestUpdate' is not initialized in the
constructor.
/libretroshare/src/rsitems/rsgxsreputationitems.h:158: warning:
Cppcheck(uninitMemberVar): Member variable
'RsGxsReputationRequestItem::mLastUpdate' is not initialized in the
constructor.
/libretroshare/src/rsitems/rsgxsreputationitems.h:55: warning:
Cppcheck(noExplicitConstructor): Class 'RsReputationItem' has a
constructor with 1 argument that is not explicit.
/libretroshare/src/rsitems/rsgxsreputationitems.h:86: warning:
Cppcheck(uninitMemberVar): Member variable
'RsGxsReputationSetItem_deprecated3::mOwnOpinion' is not initialized in
the constructor.
/libretroshare/src/rsitems/rsgxsreputationitems.h:86: warning:
Cppcheck(uninitMemberVar): Member variable
'RsGxsReputationSetItem_deprecated3::mOwnOpinionTS' is not initialized
in the constructor.
/libretroshare/src/rsitems/rsgxsreputationitems.h:86: warning:
Cppcheck(uninitMemberVar): Member variable
'RsGxsReputationSetItem_deprecated3::mIdentityFlags' is not initialized
in the constructor.