Fix CppCheck in rsmsgitems.h
/libretroshare/src/rsitems/rsmsgitems.h:99: warning:
Cppcheck(uninitMemberVar): Member variable 'RsMsgItem::msgFlags' is not
initialized in the constructor.
/libretroshare/src/rsitems/rsmsgitems.h:99: warning:
Cppcheck(uninitMemberVar): Member variable 'RsMsgItem::msgId' is not
initialized in the constructor.
/libretroshare/src/rsitems/rsmsgitems.h:99: warning:
Cppcheck(uninitMemberVar): Member variable 'RsMsgItem::sendTime' is not
initialized in the constructor.
/libretroshare/src/rsitems/rsmsgitems.h:99: warning:
Cppcheck(uninitMemberVar): Member variable 'RsMsgItem::recvTime' is not
initialized in the constructor.
/libretroshare/src/rsitems/rsmsgitems.h:131: warning:
Cppcheck(uninitMemberVar): Member variable 'RsMsgTagType::rgb_color' is
not initialized in the constructor.
/libretroshare/src/rsitems/rsmsgitems.h:131: warning:
Cppcheck(uninitMemberVar): Member variable 'RsMsgTagType::tagId' is not
initialized in the constructor.
/libretroshare/src/rsitems/rsmsgitems.h:148: warning:
Cppcheck(uninitMemberVar): Member variable 'RsMsgTags::msgId' is not
initialized in the constructor.
/libretroshare/src/rsitems/rsmsgitems.h:165: warning:
Cppcheck(uninitMemberVar): Member variable 'RsMsgSrcId::msgId' is not
initialized in the constructor.
/libretroshare/src/rsitems/rsmsgitems.h:210: warning:
Cppcheck(uninitMemberVar): Member variable 'RsMsgParentId::msgId' is not
initialized in the constructor.
/libretroshare/src/rsitems/rsmsgitems.h:210: warning:
Cppcheck(uninitMemberVar): Member variable 'RsMsgParentId::msgParentId'
is not initialized in the constructor.
/libretroshare/src/rsitems/rsmsgitems.h:86: warning:
Cppcheck(noExplicitConstructor): Class 'RsMessageItem' has a constructor
with 1 argument that is not explicit.
/libretroshare/src/rsitems/rsmsgitems.h:226: warning:
Cppcheck(noExplicitConstructor): Class 'RsMsgSerialiser' has a
constructor with 1 argument that is not explicit.