Fix CppCheck in rsnxsitems.h
/libretroshare/src/rsitems/rsnxsitems.h:117: warning:
Cppcheck(uninitMemberVar): Member variable
'RsNxsSyncGrpStatsItem::request_type' is not initialized in the
constructor.
/libretroshare/src/rsitems/rsnxsitems.h:117: warning:
Cppcheck(uninitMemberVar): Member variable
'RsNxsSyncGrpStatsItem::number_of_posts' is not initialized in the
constructor.
/libretroshare/src/rsitems/rsnxsitems.h:117: warning:
Cppcheck(uninitMemberVar): Member variable
'RsNxsSyncGrpStatsItem::last_post_TS' is not initialized in the
constructor.
/libretroshare/src/rsitems/rsnxsitems.h:371: warning:
Cppcheck(uninitMemberVar): Member variable 'RsNxsMsg::pos' is not
initialized in the constructor.
/libretroshare/src/rsitems/rsnxsitems.h:371: warning:
Cppcheck(uninitMemberVar): Member variable 'RsNxsMsg::count' is not
initialized in the constructor.
/libretroshare/src/rsitems/rsnxsitems.h:410: warning:
Cppcheck(uninitMemberVar): Member variable 'RsNxsSearchReqItem::nHops'
is not initialized in the constructor.
/libretroshare/src/rsitems/rsnxsitems.h:410: warning:
Cppcheck(uninitMemberVar): Member variable 'RsNxsSearchReqItem::token'
is not initialized in the constructor.
/libretroshare/src/rsitems/rsnxsitems.h:410: warning:
Cppcheck(uninitMemberVar): Member variable
'RsNxsSearchReqItem::expiration' is not initialized in the constructor.
/libretroshare/src/rsitems/rsnxsitems.h:431: warning:
Cppcheck(uninitMemberVar): Member variable
'RsNxsSearchResultMsgItem::token' is not initialized in the constructor.
/libretroshare/src/rsitems/rsnxsitems.h:431: warning:
Cppcheck(uninitMemberVar): Member variable
'RsNxsSearchResultMsgItem::expiration' is not initialized in the
constructor.
/libretroshare/src/rsitems/rsnxsitems.h:99: warning:
Cppcheck(noExplicitConstructor): Class 'RsNxsSyncGrpReqItem' has a
constructor with 1 argument that is not explicit.
/libretroshare/src/rsitems/rsnxsitems.h:117: warning:
Cppcheck(noExplicitConstructor): Class 'RsNxsSyncGrpStatsItem' has a
constructor with 1 argument that is not explicit.
/libretroshare/src/rsitems/rsnxsitems.h:140: warning:
Cppcheck(noExplicitConstructor): Class 'RsNxsGroupPublishKeyItem' has a
constructor with 1 argument that is not explicit.
/libretroshare/src/rsitems/rsnxsitems.h:183: warning:
Cppcheck(noExplicitConstructor): Class 'RsNxsTransacItem' has a
constructor with 1 argument that is not explicit.
/libretroshare/src/rsitems/rsnxsitems.h:211: warning:
Cppcheck(noExplicitConstructor): Class 'RsNxsSyncGrpItem' has a
constructor with 1 argument that is not explicit.
/libretroshare/src/rsitems/rsnxsitems.h:257: warning:
Cppcheck(noExplicitConstructor): Class 'RsNxsEncryptedDataItem' has a
constructor with 1 argument that is not explicit.
/libretroshare/src/rsitems/rsnxsitems.h:283: warning:
Cppcheck(noExplicitConstructor): Class 'RsNxsGrp' has a constructor with
1 argument that is not explicit.
/libretroshare/src/rsitems/rsnxsitems.h:327: warning:
Cppcheck(noExplicitConstructor): Class 'RsNxsSyncMsgReqItem' has a
constructor with 1 argument that is not explicit.
/libretroshare/src/rsitems/rsnxsitems.h:351: warning:
Cppcheck(noExplicitConstructor): Class 'RsNxsSyncMsgItem' has a
constructor with 1 argument that is not explicit.
/libretroshare/src/rsitems/rsnxsitems.h:371: warning:
Cppcheck(noExplicitConstructor): Struct 'RsNxsMsg' has a constructor
with 1 argument that is not explicit.
/libretroshare/src/rsitems/rsnxsitems.h:410: warning:
Cppcheck(noExplicitConstructor): Class 'RsNxsSearchReqItem' has a
constructor with 1 argument that is not explicit.
/libretroshare/src/rsitems/rsnxsitems.h:500: warning:
Cppcheck(noExplicitConstructor): Class 'RsNxsSerialiser' has a
constructor with 1 argument that is not explicit.
/libretroshare/src/rsitems/rsnxsitems.h:237: warning:
Cppcheck(noExplicitConstructor): Class 'RsNxsSessionKeyItem' has a
constructor with 1 argument that is not explicit.