Fix CppCheck in rsrttitems.h
/libretroshare/src/rsitems/rsrttitems.h:56: warning:
Cppcheck(uninitMemberVar): Member variable 'RsRttPingItem::mSeqNo' is
not initialized in the constructor.
/libretroshare/src/rsitems/rsrttitems.h:56: warning:
Cppcheck(uninitMemberVar): Member variable 'RsRttPingItem::mPingTS' is
not initialized in the constructor.
/libretroshare/src/rsitems/rsrttitems.h:70: warning:
Cppcheck(uninitMemberVar): Member variable 'RsRttPongItem::mSeqNo' is
not initialized in the constructor.
/libretroshare/src/rsitems/rsrttitems.h:70: warning:
Cppcheck(uninitMemberVar): Member variable 'RsRttPongItem::mPingTS' is
not initialized in the constructor.
/libretroshare/src/rsitems/rsrttitems.h:70: warning:
Cppcheck(uninitMemberVar): Member variable 'RsRttPongItem::mPongTS' is
not initialized in the constructor.
/libretroshare/src/rsitems/rsrttitems.h:46: warning:
Cppcheck(noExplicitConstructor): Class 'RsRttItem' has a constructor
with 1 argument that is not explicit.