Fix CppCheck in p3gxsreputation
/libretroshare/src/services/p3gxsreputation.cc:599: warning:
Cppcheck(unreadVariable): Variable 'totalcount' is modified but its new
value is never used.
/libretroshare/src/services/p3gxsreputation.cc:969: warning:
Cppcheck(redundantIfRemove): Redundant checking of STL container element
existence before removing it.
/libretroshare/src/services/p3gxsreputation.h:72: warning:
Cppcheck(uninitMemberVar): Member variable
'Reputation::mFriendsPositive' is not initialized in the constructor.
/libretroshare/src/services/p3gxsreputation.h:72: warning:
Cppcheck(uninitMemberVar): Member variable
'Reputation::mFriendsNegative' is not initialized in the constructor.
/libretroshare/src/services/p3gxsreputation.h:72: warning:
Cppcheck(uninitMemberVar): Member variable 'Reputation::mLastUsedTS' is
not initialized in the constructor.
/libretroshare/src/services/p3gxsreputation.h:75: warning:
Cppcheck(uninitMemberVar): Member variable
'Reputation::mFriendsPositive' is not initialized in the constructor.
/libretroshare/src/services/p3gxsreputation.h:75: warning:
Cppcheck(uninitMemberVar): Member variable
'Reputation::mFriendsNegative' is not initialized in the constructor.
/libretroshare/src/services/p3gxsreputation.h:75: warning:
Cppcheck(uninitMemberVar): Member variable 'Reputation::mLastUsedTS' is
not initialized in the constructor.
/libretroshare/src/services/p3gxsreputation.h:55: warning:
Cppcheck(noExplicitConstructor): Class 'ReputationConfig' has a
constructor with 1 argument that is not explicit.
/libretroshare/src/services/p3gxsreputation.h:75: warning:
Cppcheck(noExplicitConstructor): Class 'Reputation' has a constructor
with 1 argument that is not explicit.
/libretroshare/src/services/p3gxsreputation.h:107: warning:
Cppcheck(noExplicitConstructor): Class 'p3GxsReputation' has a
constructor with 1 argument that is not explicit.