Fix CppCheck in tcpstream
/libretroshare/src/tcponudp/tcpstream.h:83: warning:
Cppcheck(noExplicitConstructor): Class 'TcpStream' has a constructor
with 1 argument that is not explicit.
/libretroshare/src/tcponudp/tcpstream.cc:84: warning:
Cppcheck(uninitMemberVar): Member variable 'TcpStream::inData' is not
initialized in the constructor.
/libretroshare/src/tcponudp/tcpstream.cc:84: warning:
Cppcheck(uninitMemberVar): Member variable 'TcpStream::outDataRead' is
not initialized in the constructor.
/libretroshare/src/tcponudp/tcpstream.cc:84: warning:
Cppcheck(uninitMemberVar): Member variable 'TcpStream::outDataNet' is
not initialized in the constructor.
/libretroshare/src/tcponudp/tcpstream.cc:2656: warning:
Cppcheck(pointerLessThanZero): A pointer can not be negative so it is
either pointless or an error to check if it is.