Fix CppCheck in rsVOIPItems.h
/plugins/VOIP/services/rsVOIPItems.h:97: warning:
Cppcheck(uninitMemberVar): Member variable 'RsVOIPDataItem::flags' is
not initialized in the constructor.
/plugins/VOIP/services/rsVOIPItems.h:97: warning:
Cppcheck(uninitMemberVar): Member variable 'RsVOIPDataItem::data_size'
is not initialized in the constructor.
/plugins/VOIP/services/rsVOIPItems.h:97: warning:
Cppcheck(uninitMemberVar): Member variable 'RsVOIPDataItem::voip_data'
is not initialized in the constructor.
/plugins/VOIP/services/rsVOIPItems.h:129: warning:
Cppcheck(uninitMemberVar): Member variable
'RsVOIPProtocolItem::protocol' is not initialized in the constructor.
/plugins/VOIP/services/rsVOIPItems.h:129: warning:
Cppcheck(uninitMemberVar): Member variable 'RsVOIPProtocolItem::flags'
is not initialized in the constructor.
/plugins/VOIP/services/rsVOIPItems.h:143: warning:
Cppcheck(uninitMemberVar): Member variable 'RsVOIPPongItem::mSeqNo' is
not initialized in the constructor.
/plugins/VOIP/services/rsVOIPItems.h:143: warning:
Cppcheck(uninitMemberVar): Member variable 'RsVOIPPongItem::mPingTS' is
not initialized in the constructor.
/plugins/VOIP/services/rsVOIPItems.h:143: warning:
Cppcheck(uninitMemberVar): Member variable 'RsVOIPPongItem::mPongTS' is
not initialized in the constructor.
/plugins/VOIP/services/rsVOIPItems.h:73: warning:
Cppcheck(noExplicitConstructor): Class 'RsVOIPItem' has a constructor
with 1 argument that is not explicit.
/plugins/VOIP/services/rsVOIPItems.h:116: warning:
Cppcheck(uninitMemberVar): Member variable 'RsVOIPBandwidthItem::flags'
is not initialized in the constructor.
/plugins/VOIP/services/rsVOIPItems.h:116: warning:
Cppcheck(uninitMemberVar): Member variable
'RsVOIPBandwidthItem::bytes_per_sec' is not initialized in the
constructor.