Fix CppCheck in AboutWidget
/retroshare-gui/src/gui/AboutWidget.h:47: warning:
Cppcheck(noExplicitConstructor): Class 'AboutWidget' has a constructor
with 1 argument that is not explicit.
/retroshare-gui/src/gui/AboutWidget.h:155: warning:
Cppcheck(noExplicitConstructor): Class 'TBoard' has a constructor with 1
argument that is not explicit.
/retroshare-gui/src/gui/AboutWidget.h:216: warning:
Cppcheck(noExplicitConstructor): Class 'NextPieceLabel' has a
constructor with 1 argument that is not explicit.
/retroshare-gui/src/gui/AboutWidget.cpp:302: warning:
Cppcheck(uninitMemberVar): Member variable 'AWidget::mTimerId' is not
initialized in the constructor.