Fix CppCheck in RsButtonOnText
/retroshare-gui/src/gui/common/RsButtonOnText.cpp:13: warning:
Cppcheck(uninitMemberVar): Member variable 'RSButtonOnText::_textEdit'
is not initialized in the constructor.
/retroshare-gui/src/gui/common/RsButtonOnText.cpp:13: warning:
Cppcheck(uninitMemberVar): Member variable
'RSButtonOnText::_textEditViewPort' is not initialized in the
constructor.
/retroshare-gui/src/gui/common/RsButtonOnText.cpp:13: warning:
Cppcheck(uninitMemberVar): Member variable 'RSButtonOnText::_textCursor'
is not initialized in the constructor.
/retroshare-gui/src/gui/common/RsButtonOnText.cpp:21: warning:
Cppcheck(uninitMemberVar): Member variable 'RSButtonOnText::_textEdit'
is not initialized in the constructor.
/retroshare-gui/src/gui/common/RsButtonOnText.cpp:21: warning:
Cppcheck(uninitMemberVar): Member variable
'RSButtonOnText::_textEditViewPort' is not initialized in the
constructor.
/retroshare-gui/src/gui/common/RsButtonOnText.cpp:21: warning:
Cppcheck(uninitMemberVar): Member variable 'RSButtonOnText::_textCursor'
is not initialized in the constructor.
/retroshare-gui/src/gui/common/RsButtonOnText.cpp:32: warning:
Cppcheck(uninitMemberVar): Member variable 'RSButtonOnText::_textEdit'
is not initialized in the constructor.
/retroshare-gui/src/gui/common/RsButtonOnText.cpp:32: warning:
Cppcheck(uninitMemberVar): Member variable
'RSButtonOnText::_textEditViewPort' is not initialized in the
constructor.
/retroshare-gui/src/gui/common/RsButtonOnText.cpp:32: warning:
Cppcheck(uninitMemberVar): Member variable 'RSButtonOnText::_textCursor'
is not initialized in the constructor.