Fix CppCheck in grouteritems.h
/libretroshare/src/grouter/grouteritems.h:93: warning:
Cppcheck(uninitMemberVar): Member variable
'RsGRouterAbstractMsgItem::routing_id' is not initialized in the
constructor.
/libretroshare/src/grouter/grouteritems.h:93: warning:
Cppcheck(uninitMemberVar): Member variable
'RsGRouterAbstractMsgItem::service_id' is not initialized in the
constructor.
/libretroshare/src/grouter/grouteritems.h:159: warning:
Cppcheck(uninitMemberVar): Member variable
'RsGRouterTransactionChunkItem::propagation_id' is not initialized in
the constructor.
/libretroshare/src/grouter/grouteritems.h:188: warning:
Cppcheck(uninitMemberVar): Member variable
'RsGRouterTransactionAcknItem::propagation_id' is not initialized in the
constructor.