Fix CppCheck in p3bitdht_interface.cc
/libretroshare/src/dht/p3bitdht_interface.cc:366: warning:
Cppcheck(uninitMemberVar): Member variable 'RsDhtNetPeer::mPeerType' is
not initialized in the constructor.
/libretroshare/src/dht/p3bitdht_interface.cc:366: warning:
Cppcheck(uninitMemberVar): Member variable 'RsDhtNetPeer::mDhtState' is
not initialized in the constructor.
/libretroshare/src/dht/p3bitdht_interface.cc:366: warning:
Cppcheck(uninitMemberVar): Member variable
'RsDhtNetPeer::mPeerConnectState' is not initialized in the constructor.
/libretroshare/src/dht/p3bitdht_interface.cc:366: warning:
Cppcheck(uninitMemberVar): Member variable
'RsDhtNetPeer::mPeerConnectMode' is not initialized in the constructor.
/libretroshare/src/dht/p3bitdht_interface.cc:366: warning:
Cppcheck(uninitMemberVar): Member variable
'RsDhtNetPeer::mExclusiveProxyLock' is not initialized in the
constructor.
/libretroshare/src/dht/p3bitdht_interface.cc:366: warning:
Cppcheck(uninitMemberVar): Member variable 'RsDhtNetPeer::mPeerReqState'
is not initialized in the constructor.
/libretroshare/src/dht/p3bitdht_interface.cc:66: warning:
Cppcheck(stlSize): Possible inefficient checking for 'entries'
emptiness.