Fix CppCheck in p3linkmgr.cc
/libretroshare/src/pqi/p3linkmgr.cc:827: warning:
Cppcheck(unreadVariable): Variable 'out' is assigned a value that is
never used.
/libretroshare/src/pqi/p3linkmgr.cc:106: warning:
Cppcheck(uninitMemberVar): Member variable
'peerConnectState::dhtVisible' is not initialized in the constructor.
/libretroshare/src/pqi/p3linkmgr.cc:106: warning:
Cppcheck(uninitMemberVar): Member variable
'peerConnectState::actAsServer' is not initialized in the constructor.
/libretroshare/src/pqi/p3linkmgr.cc:106: warning:
Cppcheck(uninitMemberVar): Member variable 'peerConnectState::linkType'
is not initialized in the constructor.
/libretroshare/src/pqi/p3linkmgr.cc:134: warning:
Cppcheck(uninitMemberVar): Member variable
'p3LinkMgrIMPL::mAllowTunnelConnection' is not initialized in the
constructor.
/libretroshare/src/pqi/p3linkmgr.cc:2036: warning: Cppcheck(stlSize):
Possible inefficient checking for 'connAddrs' emptiness.