Fix CppCheck in pqinetstatebox.cc
/libretroshare/src/pqi/pqinetstatebox.cc:163: warning:
Cppcheck(uninitMemberVar): Member variable 'pqiNetStateBox::mStatusTS'
is not initialized in the constructor.
/libretroshare/src/pqi/pqinetstatebox.cc:163: warning:
Cppcheck(uninitMemberVar): Member variable 'pqiNetStateBox::mDhtTS' is
not initialized in the constructor.
/libretroshare/src/pqi/pqinetstatebox.cc:163: warning:
Cppcheck(uninitMemberVar): Member variable 'pqiNetStateBox::mUPnPTS' is
not initialized in the constructor.
/libretroshare/src/pqi/pqinetstatebox.cc:163: warning:
Cppcheck(uninitMemberVar): Member variable 'pqiNetStateBox::mNatPMPTS'
is not initialized in the constructor.
/libretroshare/src/pqi/pqinetstatebox.cc:163: warning:
Cppcheck(uninitMemberVar): Member variable 'pqiNetStateBox::mWebIPTS' is
not initialized in the constructor.