Fix CppCheck in rsaccounts.cc
/libretroshare/src/rsserver/rsaccounts.cc:1033: warning:
Cppcheck(stlcstrParam): Passing the result of c_str() to a function that
takes std::string as argument no. 1 is slow and redundant.
/libretroshare/src/rsserver/rsaccounts.cc:1033: warning:
Cppcheck(stlcstrParam): Passing the result of c_str() to a function that
takes std::string as argument no. 2 is slow and redundant.
/libretroshare/src/rsserver/rsaccounts.cc:1033: warning:
Cppcheck(stlcstrParam): Passing the result of c_str() to a function that
takes std::string as argument no. 5 is slow and redundant.
/libretroshare/src/rsserver/rsaccounts.cc:1033: warning:
Cppcheck(stlcstrParam): Passing the result of c_str() to a function that
takes std::string as argument no. 8 is slow and redundant.