Fix CppCheck in p3wiki.cc
/libretroshare/src/services/p3wiki.cc:45: warning:
Cppcheck(uninitMemberVar): Member variable 'p3Wiki::mAboutToken' is not
initialized in the constructor.
/libretroshare/src/services/p3wiki.cc:45: warning:
Cppcheck(uninitMemberVar): Member variable 'p3Wiki::mAboutLines' is not
initialized in the constructor.
/libretroshare/src/services/p3wiki.cc:45: warning:
Cppcheck(uninitMemberVar): Member variable 'p3Wiki::mImprovToken' is not
initialized in the constructor.
/libretroshare/src/services/p3wiki.cc:45: warning:
Cppcheck(uninitMemberVar): Member variable 'p3Wiki::mImprovLines' is not
initialized in the constructor.
/libretroshare/src/services/p3wiki.cc:45: warning:
Cppcheck(uninitMemberVar): Member variable 'p3Wiki::mMarkdownToken' is
not initialized in the constructor.
/libretroshare/src/services/p3wiki.cc:45: warning:
Cppcheck(uninitMemberVar): Member variable 'p3Wiki::mMarkdownLines' is
not initialized in the constructor.