Fix CppCheck in bdhistory.h
/libbitdht/src/bitdht/bdhistory.h:19: warning:
Cppcheck(uninitMemberVar): Member variable 'MsgRegister::msgType' is not
initialized in the constructor.
/libbitdht/src/bitdht/bdhistory.h:19: warning:
Cppcheck(uninitMemberVar): Member variable 'MsgRegister::incoming' is
not initialized in the constructor.
/libbitdht/src/bitdht/bdhistory.h:19: warning:
Cppcheck(uninitMemberVar): Member variable 'MsgRegister::aboutId' is not
initialized in the constructor.
/libbitdht/src/bitdht/bdconnection.h:66: warning:
Cppcheck(uninitMemberVar): Member variable 'bdProxyTuple::srcId' is not
initialized in the constructor.
/libbitdht/src/bitdht/bdconnection.h:66: warning:
Cppcheck(uninitMemberVar): Member variable 'bdProxyTuple::proxyId' is
not initialized in the constructor.
/libbitdht/src/bitdht/bdconnection.h:66: warning:
Cppcheck(uninitMemberVar): Member variable 'bdProxyTuple::destId' is not
initialized in the constructor.
/libbitdht/src/bitdht/bdconnection.h:171: warning:
Cppcheck(uninitMemberVar): Member variable
'bdConnectionRequest::mTarget' is not initialized in the constructor.