fix retroshare ipc
Patch by Cy

It fixes a bug, where the IPC client process sets up the shared memory, then immediately deletes the shared memory, then notifies the main GUI that the shared memory has been set up. Then the main GUI reports an error because there is no shared memory, and can't open the link that used to be in the now deleted shared memory.

So basically "$ retroshare retroshare://..." does nothing, even if the main GUI is running. At least that's what was happening for me.