Stop RSQThreadUtils::postToObject from segfaulting internally because it's broke
Some weird synchronization bug cause postToObject to segfault, when I pass the "this" pointer to a ChatWidget, which is a valid QObject. I was posting once per history line, so by accumulating all the history lines to re-add to chat, I can turn that into a single postToObject, which doesn't trigger the segfault like doing it in a loop does.

(Maybe it's bug where you segfault if you post in thread A, to thread B, then post in thread A to thread B again before the routine in thread B finishes?)