gui: chat: add auto shrink chattextedit
To shrink text edit form when it is convenient for browsing text browser.
It just shrink/unshrink the edit here and there when switching between chat rooms.

P.S. user should not resize chattextedit when it's shrinked
P.P.S. okey, there is no "unshrink" maybe, but hope it's not big deal